refactor: Remove duplicated file
