+++ src/server/routes/register.js
@@ -22,7 +22,7 @@ import * as error from '../../common/helpers/error';
 import * as handler from '../helpers/handler';
 import * as middleware from '../helpers/middleware';
 import * as propHelpers from '../../client/helpers/props';
-import * as search from '../helpers/search';
+import * as search from '../../common/helpers/search';
 import {escapeProps, generateProps} from '../helpers/props';
 import Layout from '../../client/containers/layout';
 import React from 'react';
